Update ClearPanel
보스 격파 시 2초간 터치 안되도록 수정
