+++ Assets/Scripts/DelayFunction.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DelayFunction : MonoBehaviour
+{
+    [SerializeField] Action action;
+
+    private void OnEnable()
+    {
+
+    }
+}
