+++ Assets/Scripts/DelayFunction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ae22c71be1545e478a594ebf5fea8ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
